implementations/python/requirements.txt
PyMySQL==1.1.0
SQLAlchemy==2.0.25
websocket_client==1.7.0
